requirements.txt
streamlit>=1.45.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.0
plotly>=6.0.0
pandas>=2.2.0
numpy>=1.26.0
tabulate>=0.9.0
google-genai>=1.14.0
pytz>=2024.1